+++ game_patch/graphics/d3d11/gr_d3d11_dynamic_geometry.h
@@ -21,7 +21,7 @@ namespace df::gr::d3d11
         void line(const gr::Vertex **vertices, rf::gr::Mode mode, bool is_3d);
         void line_3d(const rf::gr::Vertex& v0, const rf::gr::Vertex& v1, rf::gr::Mode mode);
         void line_2d(float x1, float y1, float x2, float y2, rf::gr::Mode mode);
-        void bitmap(int bm_handle, int x, int y, int w, int h, int sx, int sy, int sw, int sh, bool flip_x, bool flip_y, gr::Mode mode);
+        void bitmap(int bm_handle, float x, float y, float w, float h, float sx, float sy, float sw, float sh, bool flip_x, bool flip_y, gr::Mode mode);
         void flush();
 
     private:
